Refined atlas debug mode

+++ BruteForceEngine/DescriptorHeapManager.cpp
@@ -53,8 +53,8 @@ namespace BruteForce
 
 	void DescriptorHeapRange::Fill(DescriptorRange* range, size_t reg)
 	{
-		range->Init(m_Type, m_Size, static_cast<UINT>(reg));
-		range->OffsetInDescriptorsFromTableStart = m_Start;
+		range->Init(m_Type, static_cast<UINT>(m_Size), static_cast<UINT>(reg));
+		range->OffsetInDescriptorsFromTableStart = static_cast<UINT>(m_Start);
 	}
 
 }
